feat(seahaven-cli): streamline setup file and project dir path resolution
- Introduced a new function `resolve_setup_file_and_project_dir_paths` to handle the resolution of setup file and project directory paths, improving error handling and defaulting behavior.
- Refactored existing command implementations to utilize the new resolution function, enhancing code clarity and maintainability.
- Moved the `load_setup_file` function to a new `setup_yaml` module, further organizing the file processing logic.

This update enhances the Seahaven CLI's setup file handling capabilities and improves overall code organization.
