+++ seahaven-cli/src/bin/truman/cmd/run.rs
@@ -6,7 +6,10 @@ use seahaven_just::cmd::{IntoCommand, JustCmd};
 use super::common::{env_file_arg, file_arg, project_directory_arg};
 use crate::{
     deps::resolve_just_executable,
-    files::env::{load_and_merge_envs, load_setup_file_env},
+    files::{
+        env::{load_and_merge_envs, load_setup_file_env},
+        resolve_setup_file_and_project_dir_paths,
+    },
 };
 
 /// The `run` command name
@@ -49,22 +52,14 @@ pub async fn run(matches: &clap::ArgMatches) -> Result<()> {
 
     tracing::debug!("just version: {}", just_version);
 
-    // Resolve the project directory
-    let project_directory = matches
-        .get_one::<PathBuf>("project-directory")
-        .expect("Failed to get project directory");
+    let (setup_file, project_directory) = resolve_setup_file_and_project_dir_paths(
+        matches
+            .get_one::<PathBuf>("file")
+            .expect("Failed to get setup file"),
+        matches.get_one::<PathBuf>("project-directory"),
+    )?;
 
-    tracing::debug!("Project directory: {}", project_directory.display());
-
-    // Load the setup file
-    let setup_file = matches
-        .get_one::<PathBuf>("file")
-        .expect("Failed to get setup file");
-    if !setup_file.exists() {
-        return Err(anyhow::anyhow!("Setup file not found: {}", setup_file.display()).into());
-    }
-
-    let front_matter_env = load_setup_file_env(setup_file)
+    let front_matter_env = load_setup_file_env(&setup_file)
         .map_err(|err| anyhow::anyhow!("Failed to parse setup file: {err}"))?;
 
     let env = load_and_merge_envs(
